add base bottom sheet
