+++ app/src/main/java/com/innoprog/android/base/BaseBottomSheetFragment.kt
@@ -1,0 +1,29 @@
+package com.innoprog.android.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.viewbinding.ViewBinding
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+
+abstract class BaseBottomSheetFragment<T : ViewBinding> : BottomSheetDialogFragment() {
+
+    private var _binding: T? = null
+    protected val binding get() = _binding!!
+
+    abstract fun createBinding(inflater: LayoutInflater, container: ViewGroup?): T
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = createBinding(inflater, container)
+        return binding.root
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
+}
